Add markdown definition list support.
