+++ package.json
@@ -118,6 +118,7 @@
     "markdown-it": "^8.4.2",
     "markdown-it-anchor": "^5.0.2",
     "markdown-it-attrs": "^3.0.0",
+    "markdown-it-deflist": "^2.1.0",
     "mocha": "^6.2.0",
     "moment": "^2.24.0",
     "move-file": "^2.0.0",
