+++ src/site/content/en/handbook/web-dev-components/index.md
@@ -26,6 +26,7 @@ guidance about how to use them effectively.
 1. [Columns](#columns)
 1. [Code](#code)
 1. [Compare](#compare)
+1. [Definition lists](#definition-lists)
 1. [Details](#details)
 1. [Glitches](#glitches)
 1. [Images](#images)
@@ -553,6 +554,24 @@ Lorem ipsum dolor sit amet consectetur, adipisicing elit. Enim necessitatibus
 incidunt harum reprehenderit laboriosam labore consequuntur quod. Doloribus,
 deleniti! Atque aliquam facilis labore odio similique provident illo culpa
 assumenda perspiciatis.
+
+## Definition lists
+
+```md
+First Term
+: This is the definition of the first term.
+
+Second Term
+: This is one definition of the second term.
+: This is another definition of the second term.
+```
+
+First Term
+: This is the definition of the first term.
+
+Second Term
+: This is one definition of the second term.
+: This is another definition of the second term.
 
 ## Details
